move merge_imgs
