+++ vq_gan_3d/model/vqgan.py
@@ -12,7 +12,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 import torch.distributed as dist
 
-from vq_gan_3d.utils import shift_dim, adopt_weight, comp_getattr
+from vq_gan_3d.utils import shift_dim, adopt_weight, comp_getattr, merge_images
 from vq_gan_3d.model.lpips import LPIPS
 from vq_gan_3d.model.codebook import Codebook
 
@@ -99,22 +99,6 @@ class VQGAN(pl.LightningModule):
                 return vq_output['encodings']
         return h
     
-    def merge_images(self, img1, img2):
-        """
-        Merge two images by fading the last part of the first image and the first half of the second image
-        """
-        assert img1.shape[3:] == img2.shape[3:], "Images must have the same shape across Y and Z axis"
-
-        fade_length = img2.shape[2] // 2
-        fade_factor = np.linspace(0, 1, fade_length)[np.newaxis, np.newaxis, :, np.newaxis, np.newaxis]
-
-        fade1 = img1[:, :, -fade_length:, :, :] * (1 - fade_factor)
-        fade2 = img2[:, :, :fade_length, :, :] * fade_factor
-
-        fade = fade1 + fade2
-        merged = np.concatenate((img1[:, :, :-fade_length, :, :], fade, img2[:, :, fade_length:, :, :]), axis=2)
-
-        return merged
 
     def decode(self, latent, quantize=True):
         """
@@ -151,7 +135,7 @@ class VQGAN(pl.LightningModule):
             if i == 0:
                 decoded = decoded_part_np
             else:
-                decoded = self.merge_images(decoded, decoded_part_np)
+                decoded = merge_images(decoded, decoded_part_np)
 
         decoded = torch.from_numpy(decoded).to(self.device)
         return decoded
